# specs/005-user-org-service/contracts/user-org-service.openapi.yaml
openapi: 3.1.0
info:
  title: User & Organization Service API
  version: 0.1.0
  description: >
    Contracts for interactive admin API, authentication flows, policy decisions,
    budget enforcement, declarative reconciliation controls, and audit exports.
servers:
  - url: https://user-org-service.{env}.platform.internal
    description: Cluster-scoped service endpoint (dev/staging/prod)
  - url: http://localhost:8081
    description: Local development (admin API)
...
security:
  - bearerAuth: []
  - m2mToken: []
paths:
...
  /v1/orgs/{orgId}:
    get:
      tags: [Organizations]
      summary: Retrieve organization details
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgId'
      responses:
        '200':
          description: Organization fetched
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Organizations]
      summary: Update organization metadata or policies
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateOrgRequest'
      responses:
        '200':
          description: Organization updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Organization'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /v1/orgs/{orgId}/service-accounts/{serviceAccountId}/api-keys:
    post:
      tags: [APIKeys]
      summary: Issue a new API key
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - $ref: '#/components/parameters/ServiceAccountId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssueApiKeyRequest'
      responses:
        '201':
          description: API key issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IssuedApiKey'
        '429':
          description: Rotation rate limited
  /v1/orgs/{orgId}/api-keys/{apiKeyId}:
    delete:
      tags: [APIKeys]
      summary: Revoke an API key
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - $ref: '#/components/parameters/ApiKeyId'
      responses:
        '204':
          description: API key revoked
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Refresh token with dedicated scope
    m2mToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Machine-to-machine token issued to downstream services
  parameters:
    OrgId:
      name: orgId
      in: path
      required: true
      schema:
        type: string
        format: slug
...
    ServiceAccountId:
      name: serviceAccountId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ApiKeyId:
      name: apiKeyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
...
    Conflict:
      description: Resource state conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...
    Organization:
      type: object
      required: [orgId, name, slug, status]
      properties:
        orgId:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        status:
          type: string
          enum: [active, suspended, pending_delete]
        budgets:
          $ref: '#/components/schemas/BudgetStatus'
        declarative:
          $ref: '#/components/schemas/DeclarativeStatus'
        metadata:
          type: object
          additionalProperties: true
    UpdateOrgRequest:
      type: object
      properties:
        displayName:
          type: string
        status:
          type: string
          enum: [active, suspended]
        budgetPolicyId:
          type: string
          format: uuid
        declarative:
          type: object
          properties:
            enabled:
              type: boolean
            repoUrl:
              type: string
              format: uri
            branch:
              type: string
...
    IssueApiKeyRequest:
      type: object
      required: [scopes]
      properties:
        scopes:
          type: array
          items:
            type: string
        expiresInDays:
          type: integer
        annotations:
          type: object
          additionalProperties: true
    IssuedApiKey:
      type: object
      required: [apiKeyId, secret, fingerprint, status]
      properties:
        apiKeyId:
          type: string
          format: uuid
        secret:
          type: string
          description: Displayed once
        fingerprint:
          type: string
        status:
          type: string
          enum: [active, pending_rotation]
        expiresAt:
          type: string
          format: date-time
...
    BudgetStatus:
      type: object
      required: [period, spend, currency, warnThreshold, blockThreshold]
      properties:
        period:
          type: string
        spend:
          type: number
        currency:
          type: string
        warnThreshold:
          type: number
        blockThreshold:
          type: number
        status:
          type: string
          enum: [ok, warn, block]
        overridesActive:
          type: boolean
        pendingOverride:
          $ref: '#/components/schemas/BudgetOverride'
...
    BudgetOverride:
      type: object
      required: [overrideId, status]
      properties:
        overrideId:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, approved, rejected, expired]
        amount:
          type: number
        requestedBy:
          type: string
          format: uuid
        approvedBy:
          type: array
          items:
            type: string
          description: Approver user IDs
        expiresAt:
          type: string
          format: date-time
...
    DeclarativeStatus:
      type: object
      required: [mode, lastCommit, state]
      properties:
        mode:
          type: string
          enum: [enabled, disabled, paused]
        lastCommit:
          type: string
        state:
          type: string
          enum: [synced, drift_detected, failed, pending]
        lastSyncedAt:
          type: string
          format: date-time
        driftDiff:
          type: object
          additionalProperties: true
...